healer: fix issue #129
