+++ e2e-smoke/node/add.mjs
@@ -1,3 +1,1 @@
-export function add(a, b) {
-  return a + b;
-}
+export { add } from './src/add.js';
