+++ e2e-smoke/node/add.test.mjs
@@ -1,16 +1,1 @@
-import assert from 'node:assert/strict';
-import test from 'node:test';
-
-import { add } from './add.mjs';
-
-test('add returns the sum of two numbers', () => {
-  assert.equal(add(2, 3), 5);
-});
-
-test('add handles zero correctly', () => {
-  assert.equal(add(0, 0), 0);
-});
-
-test('add handles negative numbers', () => {
-  assert.equal(add(-2, 3), 1);
-});
+import './test/add.test.js';
