+++ e2e-smoke/node/package.json
@@ -4,6 +4,6 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "test": "node --test"
+    "test": "sh -c 'node --test'"
   }
 }
